+++ lib/ui/screen/map/map_screen.dart
@@ -5,7 +5,7 @@ import 'package:food_gram_app/core/model/posts.dart';
 import 'package:food_gram_app/core/utils/async_value_group.dart';
 import 'package:food_gram_app/core/utils/location.dart';
 import 'package:food_gram_app/env.dart';
-import 'package:food_gram_app/ui/component/modal_sheet/app_modal_sheet.dart';
+import 'package:food_gram_app/ui/component/modal_sheet/app_map_restaurant_modal_sheet.dart';
 import 'package:food_gram_app/ui/screen/map/maplibre_controller.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:maplibre_gl/maplibre_gl.dart';
@@ -61,7 +61,7 @@ class MapScreen extends HookConsumerWidget {
               ),
               Visibility(
                 visible: isTapPin.value,
-                child: RestaurantInfoModalSheet(post: post.value),
+                child: AppMapRestaurantModalSheet(post: post.value),
               ),
             ],
           );
