+++ lib/ui/screen/post/post_screen.dart
@@ -8,7 +8,7 @@ import 'package:food_gram_app/core/model/tag.dart';
 import 'package:food_gram_app/gen/l10n/l10n.dart';
 import 'package:food_gram_app/ui/component/app_loading.dart';
 import 'package:food_gram_app/ui/component/app_text_field.dart';
-import 'package:food_gram_app/ui/component/modal_sheet/app_modal_sheet.dart';
+import 'package:food_gram_app/ui/component/modal_sheet/app_post_image_modal_sheet.dart';
 import 'package:food_gram_app/ui/screen/post/provider/post_screen_state_provider.dart';
 import 'package:food_gram_app/utils/provider/loading.dart';
 import 'package:food_gram_app/utils/snack_bar_manager.dart';
@@ -115,7 +115,7 @@ class PostScreen extends HookConsumerWidget {
                             showModalBottomSheet(
                               context: context,
                               builder: (context) {
-                                return AppImageModalSheet(
+                                return AppPostImageModalSheet(
                                   camera: () async {
                                     await ref
                                         .read(
